pyproject.toml
[project]
name = "janus-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Harold", email = "bearljj@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "prompt-toolkit>=3.0.52",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
